+++ src/components/bluepad32/include/uni_utils.h
@@ -25,6 +25,6 @@ limitations under the License.
 // Little-endian CRC32.
 // ESP32 has its own crc32_le as well, but they don't return the same values (?).
 // It is important to use ours with the "uni_" prefix.
-uint32_t uni_crc32_le(uint32_t seed, const void* data, size_t len);
+uint32_t uni_crc32_le(uint32_t crc, const uint8_t* data, size_t len);
 
 #endif  // UNI_UTILS_H