+++ src/components/bluepad32/uni_utils.c
@@ -20,14 +20,12 @@ limitations under the License.
 
 #define CRCPOLY 0xedb88320
 
-uint32_t uni_crc32_le(uint32_t seed, const void* data, size_t len) {
-    uint32_t crc = seed;
-    const uint8_t* src = data;
+uint32_t uni_crc32_le(uint32_t crc, const uint8_t* data, size_t len) {
     uint32_t mult;
     int i;
 
     while (len--) {
-        crc ^= *src++;
+        crc ^= *data++;
         for (i = 0; i < 8; i++) {
             mult = (crc & 1) ? CRCPOLY : 0;
             crc = (crc >> 1) ^ mult;